feat: Custom Seaskipper screen redesign
