+++ common/src/main/java/com/wynntils/features/ui/CustomSeaskipperScreenFeature.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2023.
+ * Copyright © Wynntils 2023-2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.features.ui;
@@ -11,8 +11,7 @@ import com.wynntils.core.persisted.config.Category;
 import com.wynntils.core.persisted.config.Config;
 import com.wynntils.core.persisted.config.ConfigCategory;
 import com.wynntils.mc.event.ScreenOpenedEvent;
-import com.wynntils.screens.maps.SeaskipperDepartureBoardScreen;
-import com.wynntils.screens.maps.SeaskipperMapScreen;
+import com.wynntils.screens.maps.CustomSeaskipperScreen;
 import com.wynntils.utils.colors.CustomColor;
 import com.wynntils.utils.mc.McUtils;
 import com.wynntils.utils.render.type.PointerType;
@@ -22,9 +21,6 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 @ConfigCategory(Category.UI)
 public class CustomSeaskipperScreenFeature extends Feature {
     @Persisted
-    public final Config<SeaskipperScreenType> screenType = new Config<>(SeaskipperScreenType.DEPARTURE_BOARD);
-
-    @Persisted
     public final Config<PointerType> pointerType = new Config<>(PointerType.ARROW);
 
     @Persisted
@@ -33,21 +29,13 @@ public class CustomSeaskipperScreenFeature extends Feature {
     @SubscribeEvent
     public void onScreenOpen(ScreenOpenedEvent.Post event) {
         if (!Models.Container.isSeaskipper(event.getScreen().getTitle())) return;
+        if (McUtils.player().isShiftKeyDown()) return;
 
         if (!Models.Seaskipper.isProfileLoaded()) {
             McUtils.sendErrorToClient(I18n.get("feature.wynntils.customSeaskipperScreen.error.noProfile"));
             return;
         }
 
-        if (screenType.get() == SeaskipperScreenType.DEPARTURE_BOARD) {
-            McUtils.mc().setScreen(SeaskipperDepartureBoardScreen.create());
-        } else if (screenType.get() == SeaskipperScreenType.MAP) {
-            McUtils.mc().setScreen(SeaskipperMapScreen.create());
-        }
-    }
-
-    public enum SeaskipperScreenType {
-        DEPARTURE_BOARD,
-        MAP
+        McUtils.mc().setScreen(CustomSeaskipperScreen.create());
     }
 }
