+++ common/src/main/java/com/wynntils/models/seaskipper/SeaskipperModel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2023.
+ * Copyright © Wynntils 2023-2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.models.seaskipper;
@@ -17,8 +17,7 @@ import com.wynntils.mc.event.MenuEvent;
 import com.wynntils.models.items.items.gui.SeaskipperDestinationItem;
 import com.wynntils.models.seaskipper.type.SeaskipperDestination;
 import com.wynntils.models.seaskipper.type.SeaskipperDestinationProfile;
-import com.wynntils.screens.maps.SeaskipperDepartureBoardScreen;
-import com.wynntils.screens.maps.SeaskipperMapScreen;
+import com.wynntils.screens.maps.CustomSeaskipperScreen;
 import com.wynntils.services.map.pois.SeaskipperDestinationPoi;
 import com.wynntils.utils.mc.McUtils;
 import com.wynntils.utils.wynn.ContainerUtils;
@@ -88,10 +87,8 @@ public final class SeaskipperModel extends Model {
 
         // We added a new destination, reload the map
         // (This reloads the pois for every item parsed, but performance is not an issue here)
-        if (McUtils.mc().screen instanceof SeaskipperMapScreen seaskipperMapScreen) {
-            seaskipperMapScreen.reloadDestinationPois();
-        } else if (McUtils.mc().screen instanceof SeaskipperDepartureBoardScreen seaskipperDepartureBoardScreen) {
-            seaskipperDepartureBoardScreen.reloadDestinationPois();
+        if (McUtils.mc().screen instanceof CustomSeaskipperScreen customSeaskipperScreen) {
+            customSeaskipperScreen.reloadDestinationPois();
         }
     }
