+++ common/src/main/java/com/wynntils/screens/maps/widgets/SeaskipperTravelButton.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2023.
+ * Copyright © Wynntils 2023-2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.screens.maps.widgets;
@@ -7,7 +7,7 @@ package com.wynntils.screens.maps.widgets;
 import com.google.common.collect.Lists;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.wynntils.screens.base.widgets.WynntilsButton;
-import com.wynntils.screens.maps.SeaskipperDepartureBoardScreen;
+import com.wynntils.screens.maps.CustomSeaskipperScreen;
 import com.wynntils.utils.mc.McUtils;
 import com.wynntils.utils.render.RenderUtils;
 import com.wynntils.utils.render.Texture;
@@ -17,17 +17,16 @@ import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 
 public class SeaskipperTravelButton extends WynntilsButton {
-    private final SeaskipperDepartureBoardScreen departureBoardScreen;
+    private final CustomSeaskipperScreen seaskipperScreen;
 
-    public SeaskipperTravelButton(
-            int x, int y, int width, int height, SeaskipperDepartureBoardScreen departureBoardScreen) {
+    public SeaskipperTravelButton(int x, int y, int width, int height, CustomSeaskipperScreen seaskipperScreen) {
         super(x, y, width, height, Component.literal("Travel Button"));
-        this.departureBoardScreen = departureBoardScreen;
+        this.seaskipperScreen = seaskipperScreen;
     }
 
     @Override
     public void onPress() {
-        departureBoardScreen.travelToDestination();
+        seaskipperScreen.travelToDestination();
     }
 
     @Override
@@ -43,16 +42,17 @@ public class SeaskipperTravelButton extends WynntilsButton {
                 this.width,
                 this.height,
                 0,
-                departureBoardScreen.hasSelectedDestination() ? 0 : Texture.TRAVEL_BUTTON.height() / 2,
+                seaskipperScreen.getSelectedDestination() != null ? 0 : Texture.TRAVEL_BUTTON.height() / 2,
                 Texture.TRAVEL_BUTTON.width(),
                 Texture.TRAVEL_BUTTON.height() / 2,
                 Texture.TRAVEL_BUTTON.width(),
                 Texture.TRAVEL_BUTTON.height());
 
-        if (isHovered && departureBoardScreen.hasSelectedDestination()) {
-            List<Component> tooltip =
-                    List.of(Component.translatable("screens.wynntils.seaskipperMapGui.travelToDestination")
-                            .withStyle(ChatFormatting.GRAY));
+        if (isHovered && seaskipperScreen.getSelectedDestination() != null) {
+            List<Component> tooltip = List.of(Component.translatable(
+                            "screens.wynntils.customSeaskipperScreen.travelToDestination",
+                            seaskipperScreen.getSelectedDestination().getName())
+                    .withStyle(ChatFormatting.GRAY));
 
             McUtils.mc().screen.setTooltipForNextRenderPass(Lists.transform(tooltip, Component::getVisualOrderText));
         }
