+++ common/src/main/resources/assets/wynntils/lang/en_us.json
@@ -371,8 +371,6 @@
   "feature.wynntils.customSeaskipperScreen.pointerColor.name": "Pointer Color",
   "feature.wynntils.customSeaskipperScreen.pointerType.description": "What style should the main map pointer have?",
   "feature.wynntils.customSeaskipperScreen.pointerType.name": "Main Map Pointer Type",
-  "feature.wynntils.customSeaskipperScreen.screenType.description": "Which screen would you like to use for the Seaskipper UI?",
-  "feature.wynntils.customSeaskipperScreen.screenType.name": "Seaskipper UI screen",
   "feature.wynntils.customTradeMarketResultScreen.description": "Adds a custom trade market result screen where you can use Wynntils filters to filter the results.",
   "feature.wynntils.customTradeMarketResultScreen.name": "Custom Trade Market Result Screen",
   "feature.wynntils.customTradeMarketResultScreen.shiftBehaviorConfig.description": "What should happen when you press the search button while holding shift?",
@@ -1798,6 +1796,23 @@
   "screens.wynntils.containers.cancel.name": "[X] Cancel name change",
   "screens.wynntils.containers.edit.description": "Right-click to reset to default.",
   "screens.wynntils.containers.edit.name": "[>] Change current page name",
+  "screens.wynntils.customSeaskipperScreen.buyBoat.description": "Click here to buy a boat for 2²½.",
+  "screens.wynntils.customSeaskipperScreen.buyBoat.name": "Buy a boat",
+  "screens.wynntils.customSeaskipperScreen.cannotAfford": "You cannot afford to travel here.",
+  "screens.wynntils.customSeaskipperScreen.clickToGo": "Click to go here!",
+  "screens.wynntils.customSeaskipperScreen.cost": "Cost %d²",
+  "screens.wynntils.customSeaskipperScreen.destination": "%s | Level %d",
+  "screens.wynntils.customSeaskipperScreen.inaccessible": "Inaccessible",
+  "screens.wynntils.customSeaskipperScreen.level": "Level %d",
+  "screens.wynntils.customSeaskipperScreen.origin": "Origin",
+  "screens.wynntils.customSeaskipperScreen.select": "Click to select %s as your destination",
+  "screens.wynntils.customSeaskipperScreen.showBorders.description": "Click here to enable/disable territory borders.",
+  "screens.wynntils.customSeaskipperScreen.showBorders.name": "Show territory borders",
+  "screens.wynntils.customSeaskipperScreen.showInaccessibleLocations.description": "Click here to enable/disable inaccessible locations.",
+  "screens.wynntils.customSeaskipperScreen.showInaccessibleLocations.name": "Show inaccessible locations",
+  "screens.wynntils.customSeaskipperScreen.showRoutes.description": "Click here to enable/disable Seaskipper routes.",
+  "screens.wynntils.customSeaskipperScreen.showRoutes.name": "Show Seaskipper routes",
+  "screens.wynntils.customSeaskipperScreen.travelToDestination": "Click here to travel to %s",
   "screens.wynntils.gearViewer.viewStats": "View Player's Statistics",
   "screens.wynntils.guildMap.cycleDefenseFilter.description1": "Click here to cycle the territory defense filter.",
   "screens.wynntils.guildMap.cycleDefenseFilter.description2": "Hold SHIFT to filter higher, CTRL to filter lower.",
@@ -1964,15 +1979,6 @@
   "screens.wynntils.savedItems.name": "Item Record",
   "screens.wynntils.savedItems.unableToShare": "Unable to share item",
   "screens.wynntils.searchWidget.defaultSearchText": "Search...",
-  "screens.wynntils.seaskipperMapGui.buyBoat.description": "Click here to buy a boat for 2²½.",
-  "screens.wynntils.seaskipperMapGui.buyBoat.name": "Buy a boat",
-  "screens.wynntils.seaskipperMapGui.showBorders.description": "Click here to enable/disable territory borders.",
-  "screens.wynntils.seaskipperMapGui.showBorders.name": "Show territory borders",
-  "screens.wynntils.seaskipperMapGui.showInaccessibleLocations.description": "Click here to enable/disable inaccessible locations.",
-  "screens.wynntils.seaskipperMapGui.showInaccessibleLocations.name": "Show inaccessible locations",
-  "screens.wynntils.seaskipperMapGui.showRoutes.description": "Click here to enable/disable Seaskipper routes.",
-  "screens.wynntils.seaskipperMapGui.showRoutes.name": "Show Seaskipper routes",
-  "screens.wynntils.seaskipperMapGui.travelToDestination": "Click here to travel!",
   "screens.wynntils.settingsScreen.apply": "Apply",
   "screens.wynntils.settingsScreen.apply.description": "Save changes",
   "screens.wynntils.settingsScreen.booleanConfig.disabled": "Disabled",
